Removed trailing unit from variable.

+++ linux_sql/sql/ddl.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS PUBLIC.host_info(
     cpu_number INT NOT NULL,
     cpu_architecture VARCHAR NOT NULL,
     cpu_model VARCHAR NOT NULL,
-    cpu_mhz FLOAT NOT NULL,
+    cpu_mhz NUMERIC NOT NULL,
     L2_cache INT NOT NULL,
     total_mem INT NOT NULL,
     "timestamp" TIMESTAMP NOT NULL
